+++ CHANGES.md
@@ -4,7 +4,11 @@ CHANGELOG
 0.3.16.dev0    (XXXX-XX-XX)
 ---------------------------
 
+* Manage relations between layers and features (manual / intersections or distances).
+* GeoJson renderer. Now API can render .json or .geojson endpoint (or format=?geojson)
+* Partial properties patch. A patch on feature viewset keep properties not sended.
 * Add Json value search filter in FeatureViewset
+
 
 0.3.15         (2019-12-13)
 ---------------------------